Completed most of the features

+++ src/main/java/site/zeng/wukongaq/service/impl/AnswerServiceImpl.java
@@ -8,6 +8,8 @@ import site.zeng.wukongaq.entity.answer.Answer;
 import site.zeng.wukongaq.exception.AnswerException;
 import site.zeng.wukongaq.repository.answer.AnswerRepository;
 import site.zeng.wukongaq.service.AnswerService;
+import site.zeng.wukongaq.service.ProblemService;
+import site.zeng.wukongaq.service.RecommendService;
 
 import java.util.Optional;
 
@@ -18,8 +20,13 @@ import java.util.Optional;
 public class AnswerServiceImpl implements AnswerService {
     private final AnswerRepository answerRepository;
 
-    public AnswerServiceImpl(AnswerRepository answerRepository) {
+    private final ProblemService problemService;
+    private final RecommendService recommendService;
+
+    public AnswerServiceImpl(AnswerRepository answerRepository, RecommendService recommendService,ProblemService problemService) {
         this.answerRepository = answerRepository;
+        this.recommendService = recommendService;
+        this.problemService=problemService;
     }
 
 
@@ -28,6 +35,8 @@ public class AnswerServiceImpl implements AnswerService {
     public void addAnswer(Answer answer)throws AnswerException {
         try {
             answerRepository.save(answer);
+            problemService.incrementAnswerNum(answer.getPid(),1);
+            recommendService.increaseWeight(answer.getPid(),3);
         }catch (Exception e){
             throw new AnswerException(AnswerException.ANWSER_PUT_FAIL);
         }
@@ -37,7 +46,11 @@ public class AnswerServiceImpl implements AnswerService {
     @Transactional(rollbackFor = Exception.class)
     public void delAnswer(Integer uid, Integer aid) throws AnswerException {
         try {
-            answerRepository.deleteByIdAndUid(aid,uid);
+            Answer answer=answerRepository.findByIdAndUid(aid,uid);
+            if (answer!=null){
+                answerRepository.deleteByIdAndUid(aid,uid);
+                problemService.incrementAnswerNum(answer.getPid(),-1);
+            }
         }catch (Exception e){
             throw new AnswerException(AnswerException.ANWSER_DEL_FAIL);
         }
